requirements.txt
Flask==2.0.3
Flask-cors==3.0.10
Pillow==8.4.0
paramiko==2.11.0
torch~=1.12.0
numpy==1.19.5
facenet_pytorch==2.5.2
